+++ src/clickSupport/index.ts
@@ -1,0 +1,1 @@
+export { LottieClickSupportProvider } from './provider';